Split files for definition of handlers
